# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dokumentasi API Methodly",
        "description": "Dokumentasi API untuk sistem Backend Methodly.",
        "contact": {
            "email": "support@methodly.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Server API Methodly"
        }
    ],
    "paths": {
...
        "/api/v1/register": {
            "post": {
                "tags": [
                    "Autentikasi"
                ],
                "summary": "Daftar pengguna baru",
                "operationId": "6796c44e1f6079c411abbfed066c5907",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Nama lengkap pengguna",
                                        "type": "string",
                                        "example": "Budi Santoso"
                                    },
                                    "email": {
                                        "description": "Alamat email unik",
                                        "type": "string",
                                        "format": "email",
                                        "example": "budi@example.com"
                                    },
                                    "password": {
                                        "description": "Kata sandi minimal 8 karakter",
                                        "type": "string",
                                        "format": "password",
                                        "example": "rahasia123"
                                    },
                                    "password_confirmation": {
                                        "description": "Konfirmasi kata sandi",
                                        "type": "string",
                                        "format": "password",
                                        "example": "rahasia123"
                                    },
                                    "role": {
                                        "description": "Peran pengguna",
                                        "type": "string",
                                        "enum": [
                                            "student",
                                            "admin"
                                        ],
                                        "example": "student"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Pengguna berhasil didaftarkan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Pengguna berhasil didaftarkan"
                                        },
                                        "access_token": {
                                            "type": "string"
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "example": "Bearer"
                                        },
                                        "user": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Kesalahan validasi data"
                    }
                }
            }
        },
        "/api/v1/login": {
            "post": {
                "tags": [
                    "Autentikasi"
                ],
                "summary": "Masuk ke sistem",
                "operationId": "da2ca4d7a5cdec77b8b45a20373535cb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "budi@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "rahasia123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login berhasil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Login berhasil"
                                        },
                                        "access_token": {
                                            "type": "string"
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "example": "Bearer"
                                        },
                                        "user": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Kredensial tidak valid (Email belum terdaftar atau kata sandi salah)"
                    },
                    "422": {
                        "description": "Kesalahan validasi data atau akun tidak ditemukan"
                    }
                }
            }
        },
        "/api/v1/logout": {
            "post": {
                "tags": [
                    "Autentikasi"
                ],
                "summary": "Keluar dari sistem",
                "operationId": "c4eddaf1a2e990cc90e4326c0d731d41",
                "responses": {
                    "200": {
                        "description": "Berhasil keluar",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Berhasil keluar"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Tidak terautentikasi"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/me": {
            "get": {
                "tags": [
                    "Autentikasi"
                ],
                "summary": "Dapatkan profil pengguna yang sedang masuk",
                "description": "Endpoint ini memerlukan token Bearer yang valid. Pastikan Anda telah melakukan login dan memasukkan token melalui tombol 'Authorize' di atas.",
                "operationId": "47353adbf2a751f3f7deac7cc52ea517",
                "responses": {
                    "200": {
                        "description": "Data pengguna yang terautentikasi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Tidak terautentikasi"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/refresh-token": {
            "post": {
                "tags": [
                    "Autentikasi"
                ],
                "summary": "Perbarui token akses",
                "description": "Gunakan refresh_token untuk mendapatkan access_token yang baru.",
                "operationId": "2cc587dcdbd87fbe33460b0b274cea3f",
                "responses": {
                    "200": {
                        "description": "Token berhasil diperbarui. access_token baru di body, refresh_token baru di Set-Cookie.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "type": "string"
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "example": "Bearer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Refresh token tidak valid atau sudah kedaluwarsa"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Masukkan token Bearer untuk akses API",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}